pyproject.toml
[build-system]
build-backend = "pdm.backend"
requires = [ "pdm-backend" ]

[project]
name = "zeed-movs-viewer"
version = "0.0.2"
license = { file = "LICENSE" }
requires-python = ">=3.12,<3.14"
classifiers = [
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
     "jdk4py>=21.0.4.1",
    "openpyxl>=3.1.5",
    "pypdf>=4.2",
    "tabula-py[jpype]>=2.9.3",
   "pyside6-essentials>=6.10",
    "pythonqwt>=0.12.5",
    "selenium>=4.22",
    "zeed-guilib>=0.0.10",
]
urls.Homepage = "https://github.com/ZeeD/movs-viewer"
urls.Repository = "https://github.com/ZeeD/movs-viewer.git"
scripts.movs-merger = "movsmerger.movsmerger:main"
scripts.movs-validator = "movsvalidator.movsvalidator:main"
scripts.movs-inflazione = "movsinflazione:main"
gui-scripts.movs-viewer = "movsviewer.mainui:main"

[dependency-groups]
dev = [
    "coverage>=7.6.1",
    "mypy>=1.11.2",
    "pyproject-fmt>=2.5",
    "ruff>=0.6.9",
    "pandas-stubs>=2.2.2.240603",
    "types-openpyxl>=3.1.5.20240918",
]

[tool.pdm]
distribution = true
plugins = [ "pdm-bump" ]

[tool.pdm.build]
excludes = [ "tests", "stubs", "demo" ]

[tool.pdm.scripts]
_.env = { PYTHONPATH = "src:tests:stubs:demo" }
lint = { composite = [
    "pyproject-fmt pyproject.toml",
    "ruff format {args:.}",
    "ruff check {args:.}",
    "mypy {args:.}",
], keep_going = true }
test = { composite = [
    "coverage run -m unittest discover --verbose --locals --failfast --catch --start-directory tests {args}",
    "coverage report",
    "coverage erase",
] }
movs-merger = "movs-merger"
movs-validator = "movs-validator"
movs-viewer = "movs-viewer"
movs-inflazione = "movs-inflazione"
start = "movs-viewer" # "main" one
bump_and_publish = { composite = [
    "pdm bump patch --commit --tag",
    "pdm publish",
    "git push",
    "git push --tags",
] }

[tool.ruff]
line-length = 80
src = [ "demo", "src", "stubs", "tests" ]
fix = true
format.quote-style = "single"
format.skip-magic-trailing-comma = true
format.docstring-code-format = true
lint.select = [ "ALL" ]
lint.ignore = [
    "ANN1",
    "COM812",
    "D1",
    "D203",
    "D213",
    "ISC001",
    "PT009",
    "PT027",
    "Q001",
    "Q002",
    "TD002",
    "TD003",
]
lint.extend-per-file-ignores."stubs/**/*.pyi" = [
    "FBT001",
    "FBT002",
    "N802",
    "N803",
    "N815",
    "PLR0913",
]
lint.flake8-annotations.allow-star-arg-any = true
lint.flake8-quotes.docstring-quotes = "single"
lint.flake8-quotes.inline-quotes = "single"
lint.flake8-quotes.multiline-quotes = "single"
lint.flake8-type-checking.quote-annotations = true
lint.isort.force-single-line = true

[tool.pyproject-fmt]
column_width = 80
indent = 4

[tool.coverage.run]
branch = true
source = [ "src", "tests" ]

[tool.coverage.report]
show_missing = true
sort = "Cover"
skip_empty = true

[tool.mypy]
mypy_path = [ "src", "tests", "stubs", "demo" ]
strict = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [ "movsviewer.automation" ]
disable_error_code = "no-untyped-call"
